[core] bootstrap-namealias without register decl
to fix with gperf-3.2 and g++ compilation. changed signature to size_t.
gperf-3.2 came up with a lot more style violations, but didn't fix register yet.
c_parens.t needed one more adjustment for #elif not being a function.

+++ include/parrot/namealias.h
@@ -17,7 +17,7 @@ gperf generated icu control character namealias lookup
 struct Parrot_namealias { int name; const INTVAL codepoint; };
 
 const struct Parrot_namealias *
-Parrot_namealias_lookup(const char *str, unsigned int len);
+Parrot_namealias_lookup(const char *str, size_t len);
 
 #endif /* PARROT_IN_CORE */
 #endif /* PARROT_NAMEALIAS_H_GUARD */
